leave core lib unchanged and move everything to samples/tools
