+++ samples/tools/finetuning/finetuning/__init__.py
@@ -1,3 +1,3 @@
-from .conversable_agent_update_model import update_model
+from .update_model import update_model
 
 __all__ = ["update_model"]
