+++ samples/tools/finetuning/tests/test_conversable_agent_update_model.py
@@ -1,28 +1,25 @@
 import pytest
 from autogen import AssistantAgent, UserProxyAgent
-
 import sys
+import os
 
 sys.path.append("samples/tools/finetuning")
 
 from finetuning import update_model  # noqa: E402
 from typing import Dict  # noqa: E402
 
-try:
-    from openai import OpenAI
-except ImportError:
-    skip = True
-else:
-    skip = False
+sys.path.append("test")
+from conftest import skip_openai  # noqa: E402
 
 TEST_CUSTOM_RESPONSE = "This is a custom response."
 TEST_LOCAL_MODEL_NAME = "local_model_name"
 
 
+@pytest.mark.skipif(skip_openai, reason="requested to skip openai tests")
 def test_custom_model_client():
     TEST_LOSS = 0.5
 
-    class CustomModel:
+    class UpdatableCustomModel:
         def __init__(self, config: Dict):
             self.model = config["model"]
             self.model_name = config["model"]
@@ -55,7 +52,7 @@ def test_custom_model_client():
         def update_model(self, preference_data, messages, **kwargs):
             return {"loss": TEST_LOSS}
 
-    config_list = [{"model": TEST_LOCAL_MODEL_NAME, "model_client_cls": "CustomModel"}]
+    config_list = [{"model": TEST_LOCAL_MODEL_NAME, "model_client_cls": "UpdatableCustomModel"}]
 
     assistant = AssistantAgent(
         "assistant",
@@ -63,7 +60,7 @@ def test_custom_model_client():
         human_input_mode="NEVER",
         llm_config={"config_list": config_list},
     )
-    assistant.register_model_client(model_client_cls=CustomModel)
+    assistant.register_model_client(model_client_cls=UpdatableCustomModel)
     user_proxy = UserProxyAgent(
         "user_proxy",
         human_input_mode="NEVER",
@@ -81,6 +78,7 @@ def test_custom_model_client():
     assert update_model_stats["update_stats"]["loss"] == TEST_LOSS
 
 
+@pytest.mark.skipif(skip_openai, reason="requested to skip openai tests")
 def test_update_model_without_client_raises_error():
     assistant = AssistantAgent(
         "assistant",
@@ -104,8 +102,9 @@ def test_update_model_without_client_raises_error():
         update_model(assistant, [], user_proxy)
 
 
+@pytest.mark.skipif(skip_openai, reason="requested to skip openai tests")
 def test_custom_model_update_func_missing_raises_error():
-    class CustomModel:
+    class UpdatableCustomModel:
         def __init__(self, config: Dict):
             self.model = config["model"]
             self.model_name = config["model"]
@@ -135,7 +134,7 @@ def test_custom_model_update_func_missing_raises_error():
         def get_usage(response) -> Dict:
             return {}
 
-    config_list = [{"model": TEST_LOCAL_MODEL_NAME, "model_client_cls": "CustomModel"}]
+    config_list = [{"model": TEST_LOCAL_MODEL_NAME, "model_client_cls": "UpdatableCustomModel"}]
 
     assistant = AssistantAgent(
         "assistant",
@@ -143,7 +142,7 @@ def test_custom_model_update_func_missing_raises_error():
         human_input_mode="NEVER",
         llm_config={"config_list": config_list},
     )
-    assistant.register_model_client(model_client_cls=CustomModel)
+    assistant.register_model_client(model_client_cls=UpdatableCustomModel)
     user_proxy = UserProxyAgent(
         "user_proxy",
         human_input_mode="NEVER",
@@ -161,8 +160,9 @@ def test_custom_model_update_func_missing_raises_error():
         update_model(assistant, [], user_proxy)
 
 
+@pytest.mark.skipif(skip_openai, reason="requested to skip openai tests")
 def test_multiple_model_clients_raises_error():
-    class CustomModel:
+    class UpdatableCustomModel:
         def __init__(self, config: Dict):
             self.model = config["model"]
             self.model_name = config["model"]
@@ -196,8 +196,8 @@ def test_multiple_model_clients_raises_error():
             return {}
 
     config_list = [
-        {"model": TEST_LOCAL_MODEL_NAME, "model_client_cls": "CustomModel"},
-        {"model": TEST_LOCAL_MODEL_NAME, "model_client_cls": "CustomModel"},
+        {"model": TEST_LOCAL_MODEL_NAME, "model_client_cls": "UpdatableCustomModel"},
+        {"model": TEST_LOCAL_MODEL_NAME, "model_client_cls": "UpdatableCustomModel"},
     ]
 
     assistant = AssistantAgent(
@@ -206,8 +206,8 @@ def test_multiple_model_clients_raises_error():
         human_input_mode="NEVER",
         llm_config={"config_list": config_list},
     )
-    assistant.register_model_client(model_client_cls=CustomModel)
-    assistant.register_model_client(model_client_cls=CustomModel)
+    assistant.register_model_client(model_client_cls=UpdatableCustomModel)
+    assistant.register_model_client(model_client_cls=UpdatableCustomModel)
     user_proxy = UserProxyAgent(
         "user_proxy",
         human_input_mode="NEVER",
